#10 [Feat] ESP32 안전장비 센서 및 SOS 드론 대응 구현 (#11)
* Feat: ESP32 FSR 및 SOS 센서 계약 구현
* Feat: LV 단계별 알림 부저 및 드론 출동 정책 구현
* Feat: 관리자 확인 후 720p 드론 영상 송출 구현
* Infra: Flyway 스키마 및 로컬 시드 데이터 구성
* Docs: MVP API 명세 및 로컬 DB 실행 문서 작성
* Test: MVP 위험 대응 및 Flyway 시드 통합 검증

+++ src/main/java/com/worksafe/backend/domain/sensor/dto/response/SensorLogResponse.java
@@ -29,7 +29,7 @@ public record SensorLogResponse(
         Double latitude,
         Double longitude,
         Double speed,
-        Double pressureValue,
+        Integer pressureValue,
         Double lidarFrontLeft,
         Double lidarFrontRight,
         Double lidarBackLeft,
